+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - `Lexer.lex` add optional parameter `expect` to limit the output token type or/and token content.
   - Add `Lexer.trimStart`.
   - Add `Lexer.take`.
+  - Add `Lexer.digested`.
   - Rename `Lexer.hasError` to `Lexer.hasErrors`.
 - Parser
   - If ASTNode has no error, the error field will be `undefined`.